поворот

+++ src/main/java/functions/test/MyPoint.java
@@ -1,11 +1,11 @@
 package functions.test;
 
 public class MyPoint {
-    public int x;
-    public int y;
-    public int z;
+    public double x;
+    public double y;
+    public double z;
 
-    public MyPoint(int x, int y, int z) {
+    public MyPoint(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
